Initial Windows OS commit

+++ src/Renderers/renderer.cpp
@@ -25,10 +25,10 @@ MatrixXd Backward_Raytracing::render(Scene scene, int spp)  {
     MatrixXd imgBuffer(resy * resx, 3);
 
     // calculate height of image plane
-    float h_img_plane = tan(fov/180.0 * M_PI/2.0);
+    double h_img_plane = tan(fov/180.0 * M_PI/2.0);
 
     // calculate step size required to travel to adjacent pixel (L/R/U/D)
-    float del_dir = h_img_plane / (resy/2.0);
+    double del_dir = h_img_plane / (resy/2.0);
 
     // calculate basis of plane
     RowVector3d i_dir = cam.dir - cam.eye; i_dir.normalize();  // normal of plane
@@ -40,8 +40,8 @@ MatrixXd Backward_Raytracing::render(Scene scene, int spp)  {
     RowVector3d del_v = i_v * del_dir;   // vector used to travel in vertical direction on image plane
     RowVector3d del_h = i_h * del_dir;  // vector used to travel in horizontal direction on image plane
 
-    float seed_v = -resy/2.0;
-    float seed_h = -resx/2.0;
+    double seed_v = -resy/2.0;
+    double seed_h = -resx/2.0;
     RowVector3d img_plane_initialPt = img_plane_centerPt - del_v * seed_v - del_h * seed_h; // initial point on plane (0,0)
     
     // iterate over all image pixels, generate a ray, and trace it
@@ -67,10 +67,10 @@ static RowVector3d trace(Scene scene, Ray r) {
 
     if (intersection[0] > 0) {
         if (scene.lights.size() == 0) {
-            Frame frame = scene.geometry[intersection[1]]._p->getFrame(r.o + r.d * intersection[2]);
+            Frame frame = scene.geometry[(int)intersection[1]]._p->getFrame(r.o + r.d * intersection[2]);
             return RowVector3d(1, 1, 1) * frame.cosTheta(frame.toLocal(-r.d));
         }
-        return shade(scene, r.o + r.d * intersection[2], -r.d, scene.geometry[intersection[1]]);
+        return shade(scene, r.o + r.d * intersection[2], -r.d, scene.geometry[(int)intersection[1]]);
     }
     return RowVector3d(0,0,0);
 }
@@ -117,7 +117,7 @@ static int intersect_shadowRay(Scene scene, Ray r) {
         a hit is found
     **/
 
-    int n = scene.geometry.size();
+    size_t n = scene.geometry.size();
     double inf, t, d;
     inf = t = r.max;
     int id = -1;
@@ -146,7 +146,7 @@ static RowVector3d intersect(Scene scene, Ray r) {
         a hit is found
     **/
 
-    int n = scene.geometry.size();
+    size_t n = scene.geometry.size();
     double inf, t, d;
     inf = t = r.max;
     int id = -1;
